slvm
Module fxhasher

Fast, non-cryptographic hash used by rustc and Firefox.

Example

use rustc_hash::FxHashMap;
let mut map: FxHashMap<u32, u32> = FxHashMap::default();
map.insert(22, 44);

Structs

FxHasher

A speedy hash algorithm for use within rustc. The hashmap in liballoc by default uses SipHash which isn’t quite as speedy as we want. In the compiler we’re not really worried about DOS attempts, so we use a fast non-cryptographic hash.

Constants

K 🔒

Type Aliases

FxHashMap

Type alias for a hashmap using the fx hash algorithm.

FxHashSet

Type alias for a hashmap using the fx hash algorithm.