slvm::heap::vm_hashmap

Struct VMHashMap

pub struct VMHashMap {
    map: HashMap<ValHash, Value, IdHasher>,
}

Wrapper class for a HashMap<Value, Value>. We need this because we want String and StringConst to hash to the same value (what a script user will expect) and that requires a VM so can not just implement Hash on Value (will not have access to a VM).

Fields
map: HashMap<ValHash, Value, IdHasher>
Implementations

impl VMHashMap

pub fn new() -> Self

Create a new empty HashMap.

pub fn with_capacity(cap: usize) -> Self

Create a new empty HashMap with an initial capacity.

pub fn get<ENV>(&self, vm: &GVm<ENV>, key: Value) -> Option<Value>

Get the value at key, requires the current VM for hashing.

pub fn insert<ENV>( &mut self, vm: &GVm<ENV>, key: Value, val: Value, ) -> Option<Value>

Insert the value at key, requires the current VM for hashing. Returns the old value at key if it exists (None otherwise).

pub fn insert_id(&mut self, id: ValHash, val: Value) -> Option<Value>

Insert val at the key id provided. This allows calling code to pre-generate the ValHash. This is a borrow checker workaround.

pub fn len(&self) -> usize

Number of items in the HashMap.

pub fn is_empty(&self) -> bool

Is this HashMap empty?

pub fn contains_key<ENV>(&self, vm: &GVm<ENV>, key: Value) -> bool

Does this HashMap contain key?

pub fn clear(&mut self)

Clear (remove all key/values) from the HashMap.

pub fn remove<ENV>(&mut self, vm: &GVm<ENV>, key: Value) -> Option<Value>

Remove key from the HashMap. Return the old value if it existed (None otherwise).

pub fn remove_id(&mut self, id: ValHash) -> Option<Value>

Remove the key from HashMap (like remove) except caller pre-generates the ValHash. Used to work around the borrow checker.

pub fn keys(&self) -> VMMapKeys<'_>

Returns an iterator over all the keys in the HashMap.

pub fn iter(&self) -> VMHashMapIter<'_>

Return an iterator over all the (key, value) pairs in the HashMap.