Crate slvm

Re-exports

pub use crate::opcodes::*;

pub use crate::error::*;

pub use crate::value::*;

pub use crate::heap::*;

pub use crate::chunk::*;

pub use crate::vm::*;

pub use crate::interner::*;

pub use crate::fxhasher::*;

Modules

chunk

error

float

This module controls access to two different floating point implementations (F56 and F32Wrap). F32Wrap is implemented in this file, and F56 is implemented in its own file. The purpose of this module is to allow the Value module to switch between the two implementations

fxhasher

Fast, non-cryptographic hash used by rustc and Firefox.

heap

interner

opcodes

value

vm

Macros

clear_bit

decode1

decode2

decode3

decode_u8

decode_u8_enum

decode_u16

decode_u32

decode_u16_enum

decode_u32_enum

get_code

get_code_at

inc_ip

is_bit_set

mov_register

set_bit

set_register