slvm::heap

Struct Heap

pub struct Heap {
    objects: Storage<Object>,
    callframes: Storage<CallFrame>,
    continuations: Storage<Continuation>,
    errors: Storage<Error>,
    pairs: Storage<(Value, Value)>,
    values: Storage<Value>,
    ios: Storage<HeapIo>,
    props: Option<FxHashMap<Value, Arc<FxHashMap<Interned, Value>>>>,
    greys: Vec<Value>,
    paused: u32,
}

Fields
objects: Storage<Object>callframes: Storage<CallFrame>continuations: Storage<Continuation>errors: Storage<Error>pairs: Storage<(Value, Value)>values: Storage<Value>ios: Storage<HeapIo>props: Option<FxHashMap<Value, Arc<FxHashMap<Interned, Value>>>>greys: Vec<Value>paused: u32
Implementations

impl Heap

pub fn new() -> Self

fn props(&self) -> &FxHashMap<Value, Arc<FxHashMap<Interned, Value>>>

fn props_mut( &mut self, ) -> &mut FxHashMap<Value, Arc<FxHashMap<Interned, Value>>>

pub fn sizeof_object() -> usize

pub fn pause_gc(&mut self)

Pause garbage collection. Each pause_gc must have an unpause_gc before GC resumes (it is a counter that must be 0).

pub fn unpause_gc(&mut self)

UnPause garbage collection. Each pause_gc must have an unpause_gc before GC resumes (it is a counter that must be 0).

pub fn set_grow_factor(&mut self, grow_factor: f64)

fn alloc<MarkFunc>( &mut self, obj: Object, flags: u8, mark_roots: MarkFunc, ) -> Handle
where MarkFunc: FnMut(&mut Heap) -> VMResult<()>,

pub fn alloc_pair<MarkFunc>( &mut self, car: Value, cdr: Value, mutable: MutState, mark_roots: MarkFunc, ) -> Value
where MarkFunc: FnMut(&mut Heap) -> VMResult<()>,

pub fn alloc_string<MarkFunc>( &mut self, s: String, mutable: MutState, mark_roots: MarkFunc, ) -> Value
where MarkFunc: FnMut(&mut Heap) -> VMResult<()>,

pub fn alloc_vector<MarkFunc>( &mut self, v: Vec<Value>, mutable: MutState, mark_roots: MarkFunc, ) -> Value
where MarkFunc: FnMut(&mut Heap) -> VMResult<()>,

pub fn alloc_map<MarkFunc>( &mut self, map: VMHashMap, mutable: MutState, mark_roots: MarkFunc, ) -> Value
where MarkFunc: FnMut(&mut Heap) -> VMResult<()>,

pub fn alloc_bytes<MarkFunc>( &mut self, v: Vec<u8>, mutable: MutState, mark_roots: MarkFunc, ) -> Value
where MarkFunc: FnMut(&mut Heap) -> VMResult<()>,

pub fn alloc_lambda<MarkFunc>( &mut self, l: Arc<Chunk>, mark_roots: MarkFunc, ) -> Value
where MarkFunc: FnMut(&mut Heap) -> VMResult<()>,

pub fn alloc_closure<MarkFunc>( &mut self, l: Arc<Chunk>, v: Vec<Handle>, mark_roots: MarkFunc, ) -> Value
where MarkFunc: FnMut(&mut Heap) -> VMResult<()>,

pub fn alloc_continuation<MarkFunc>( &mut self, k: Continuation, mark_roots: MarkFunc, ) -> Value
where MarkFunc: FnMut(&mut Heap) -> VMResult<()>,

pub fn alloc_callframe<MarkFunc>( &mut self, frame: CallFrame, mark_roots: MarkFunc, ) -> Value
where MarkFunc: FnMut(&mut Heap) -> VMResult<()>,

pub fn alloc_value<MarkFunc>( &mut self, val: Value, mutable: MutState, mark_roots: MarkFunc, ) -> Value
where MarkFunc: FnMut(&mut Heap) -> VMResult<()>,

pub fn alloc_error<MarkFunc>( &mut self, error: Error, mutable: MutState, mark_roots: MarkFunc, ) -> Value
where MarkFunc: FnMut(&mut Heap) -> VMResult<()>,

pub fn alloc_io<MarkFunc>( &mut self, io: HeapIo, mutable: MutState, mark_roots: MarkFunc, ) -> Value
where MarkFunc: FnMut(&mut Heap) -> VMResult<()>,

pub fn get_string(&self, handle: Handle) -> &str

pub fn get_string_mut(&mut self, handle: Handle) -> VMResult<&mut String>

pub fn get_vector(&self, handle: Handle) -> &[Value]

pub fn get_vector_mut(&mut self, handle: Handle) -> VMResult<&mut Vec<Value>>

pub fn get_map(&self, handle: Handle) -> &VMHashMap

pub fn get_map_mut(&mut self, handle: Handle) -> VMResult<&mut VMHashMap>

pub fn get_bytes(&self, handle: Handle) -> &[u8]

pub fn get_pair(&self, handle: Handle) -> (Value, Value)

pub fn get_pair_mut( &mut self, handle: Handle, ) -> VMResult<(&mut Value, &mut Value)>

pub fn get_pair_mut_override( &mut self, handle: Handle, ) -> (&mut Value, &mut Value)

pub fn get_lambda(&self, handle: Handle) -> Arc<Chunk>

pub fn get_closure(&self, handle: Handle) -> (Arc<Chunk>, &[Handle])

pub fn get_closure_captures(&self, handle: Handle) -> &[Handle]

pub fn get_continuation(&self, handle: Handle) -> &Continuation

pub fn get_callframe(&self, handle: Handle) -> &CallFrame

pub fn get_value(&self, handle: Handle) -> Value

pub fn get_value_mut(&mut self, handle: Handle) -> &mut Value

pub fn get_error(&self, handle: Handle) -> Error

pub fn get_io(&self, handle: Handle) -> &HeapIo

pub fn is_live(&self, val: Value) -> bool

If val is on the heap is it still alive after GC Return true if val is not a heap object.

pub fn immutable(&mut self, val: Value)

pub fn sticky(&mut self, val: Value)

pub fn unsticky(&mut self, val: Value)

pub fn is_traced_and_set(&mut self, val: Value) -> bool

pub fn mark(&mut self, value: Value)

fn mark_trace(&mut self, val: Value)

fn mark_chunk(&mut self, chunk: &Chunk)

pub fn mark_call_frame(&mut self, call_frame: &CallFrame)

fn trace_object(&mut self, obj: &Object)

fn trace(&mut self, val: Value)

fn initial_mark_call_frame(frame: &CallFrame, greys: &mut Vec<Value>)

fn collect<MarkFunc>(&mut self, mark_roots: MarkFunc)
where MarkFunc: FnMut(&mut Heap) -> VMResult<()>,

pub fn live_objects(&self) -> usize

pub fn get_property(&self, value: Value, prop: Interned) -> Option<Value>

pub fn set_property(&mut self, key_value: Value, prop: Interned, value: Value)