slvm::value

Enum Value

pub enum Value {

    Byte(u8),
    Int([u8; 7]),
    Float(F56),
    CodePoint(char),
    CharCluster(u8, [u8; 6]),
    CharClusterLong(Handle),
    Symbol(Interned),
    Keyword(Interned),
    StringConst(Interned),
    Special(Interned),
    Builtin(u32),
    True,
    False,
    Nil,
    Undefined,
    String(Handle),
    Vector(Handle),
    Map(Handle),
    Bytes(Handle),
    Pair(Handle),
    List(Handle, u16),
    Lambda(Handle),
    Closure(Handle),
    Continuation(Handle),
    CallFrame(Handle),
    Value(Handle),
    Error(Handle),
    Io(Handle),

}

Variants

Byte(u8)

Int([u8; 7])

Store a 7 byte int (i56…).

Float(F56)

Replace with float::F32Wrap if desired

CodePoint(char)

CharCluster(u8, [u8; 6])

CharClusterLong(Handle)

Handle points to a String on the heap.

Symbol(Interned)

Keyword(Interned)

StringConst(Interned)

Special(Interned)

Intended for symbols that are compiled.

Builtin(u32)

True

False

Nil

Undefined

Value is a placeholder only, should never appear at runtime. Can be instantiated at times as a placeholder but should always be overwritten.

String(Handle)

Vector(Handle)

Map(Handle)

Bytes(Handle)

Pair(Handle)

List(Handle, u16)

Lambda(Handle)

Closure(Handle)

Continuation(Handle)

CallFrame(Handle)

Value(Handle)

Error(Handle)

Io(Handle)

Implementations

impl Value

pub fn new() -> Self

pub fn unref<ENV>(self, vm: &GVm<ENV>) -> Value

pub fn get_symbol(&self) -> Option<Interned>

pub fn is_symbol(&self, sym: Interned) -> bool

pub fn is_indirect(&self) -> bool

pub fn is_nil(&self) -> bool

pub fn is_undef(&self) -> bool

pub fn is_true(&self) -> bool

pub fn is_truthy(&self) -> bool

pub fn is_false(&self) -> bool

pub fn is_falsey(&self) -> bool

pub fn is_int(&self) -> bool

pub fn is_number(&self) -> bool

pub fn is_float(&self) -> bool

pub fn not(&self) -> Value

pub fn get_int<ENV>(&self, vm: &GVm<ENV>) -> VMResult<i64>

pub fn get_float<ENV>(&self, _vm: &GVm<ENV>) -> VMResult<f64>

pub fn get_string<'vm, ENV>(&self, vm: &'vm GVm<ENV>) -> VMResult<&'vm str>

pub fn get_handle(&self) -> Option<Handle>

pub fn get_pair<ENV>(&self, vm: &GVm<ENV>) -> Option<(Value, Value)>

pub fn iter<'vm, ENV>( &self, vm: &'vm GVm<ENV>, ) -> Box<dyn Iterator<Item = Value> + 'vm>

Returns an iterator over all values in the list, vector, pair. If the item is not one of the above it returns an empty iterator.

pub fn iter_all<'vm, ENV>( &self, vm: &'vm GVm<ENV>, ) -> Box<dyn Iterator<Item = Value> + 'vm>

Returns an iterator over all values in the list, vector, pair. Returns an empty iterator if the value is nil. If the item is not one of the above it returns a once iter of the value.

Particularly useful if iterating over a Vec<Value> that contains a heterogeneous mix of List(s), Vector(s), Pair(s), and potentially other values, rather than returning an empty list for any non-iterable thing, return a once iter so the item is not missed.

pub fn display_value<ENV>(&self, vm: &GVm<ENV>) -> String

pub fn pretty_value<ENV>(&self, vm: &GVm<ENV>) -> String

pub fn value_type<ENV>(&self, vm: &GVm<ENV>) -> ValueType

Map a Value to a ValueType which can be written to a debug string that refers to the Slosh types and does not require passing in a GVm to do so.

pub fn display_type<ENV>(&self, vm: &GVm<ENV>) -> &'static str

pub fn is_proper_list<ENV>(&self, vm: &GVm<ENV>) -> bool

pub fn get_hash<ENV>(&self, vm: &GVm<ENV>) -> u64

Get the hash of a value making sure that strings (whether const or allocated) hash to the same value. Note: This use the FxHasher which means it is fast and stable but is not randomized to be hardened against external key attacks.